Rename module to package-semantic-analyzer
