+++ misc/compiler-plugins/modules/package-semantic-analyzer/build.gradle
@@ -17,7 +17,7 @@
  
 apply from: "$rootDir/gradle/javaProject.gradle"
 
-description = 'Ballerina Package Analyzer'
+description = 'Ballerina Package Semantic Analyzer'
 version = '1.0.0'
 
 dependencies {
@@ -26,7 +26,7 @@ dependencies {
     implementation project(':toml-parser')
 }
 
-ext.moduleName = 'org.ballerinalang.pkg.toml.semntic.analyzer'
+ext.moduleName = 'org.ballerinalang.pkgsemanticanalyzer'
 
 compileJava {
     doFirst {
