+++ misc/compiler-plugins/modules/package-semantic-analyzer/spotbugs-exclude.xml
@@ -18,6 +18,6 @@
 
 <FindBugsFilter>
     <Match>
-        <Package name="org.ballerinalang.pkg.toml.semntic.analyzer" />
+        <Package name="org.ballerinalang.pkgsemanticanalyzer" />
     </Match>
 </FindBugsFilter>
