+++ misc/compiler-plugins/modules/package-semantic-analyzer/src/main/java/module-info.java
@@ -2,5 +2,5 @@ module org.ballerinalang.pkg.toml.semntic.analyzer {
     requires io.ballerina.lang;
     requires io.ballerina.tools.api;
     requires io.ballerina.toml;
-    exports org.ballerinalang.pkg.toml.semntic.analyzer;
+    exports org.ballerinalang.pkgsemanticanalyzer;
 }
