+++ misc/compiler-plugins/modules/package-semantic-analyzer/src/main/resources/META-INF/services/io.ballerina.projects.plugins.CompilerPlugin
@@ -1,0 +1,1 @@
+org.ballerinalang.pkgsemanticanalyzer.PackageSemanticAnalyzerPlugin
