--- misc/compiler-plugins/modules/pkg-toml-semantic-analyzer/src/main/resources/META-INF/services/io.ballerina.projects.plugins.CompilerPlugin
@@ -1,1 +1,0 @@
-org.ballerinalang.pkg.toml.semntic.analyzer.PackageTomlSemanticAnalyzerPlugin
