+++ settings.gradle
@@ -99,7 +99,7 @@ include(':build-config:checkstyle')
 include(':debug-adapter:debug-adapter-core')
 include(':debug-adapter:debug-adapter-cli')
 include(':debug-adapter:debug-adapter-runtime')
-include(':compiler-plugins:pkg-toml-semantic-analyzer')
+include(':compiler-plugins:package-semantic-analyzer')
 include(':formatter:formatter-core')
 include(':formatter:formatter-cli')
 include(':ls-extensions:json-to-record-converter')
@@ -211,7 +211,7 @@ project(':bir-spec').projectDir = file('docs/bir-spec')
 project(':debug-adapter:debug-adapter-core').projectDir = file('misc/debug-adapter/modules/debug-adapter-core')
 project(':debug-adapter:debug-adapter-cli').projectDir = file('misc/debug-adapter/modules/debug-adapter-cli')
 project(':debug-adapter:debug-adapter-runtime').projectDir = file('misc/debug-adapter/modules/debug-adapter-runtime')
-project(':compiler-plugins:pkg-toml-semantic-analyzer').projectDir = file('misc/compiler-plugins/modules/pkg-toml-semantic-analyzer')
+project(':compiler-plugins:package-semantic-analyzer').projectDir = file('misc/compiler-plugins/modules/package-semantic-analyzer')
 project(':formatter:formatter-core').projectDir = file('misc/formatter/modules/formatter-core')
 project(':formatter:formatter-cli').projectDir = file('misc/formatter/modules/formatter-cli')
 project(':ls-extensions:json-to-record-converter').projectDir = file('misc/ls-extensions/modules/json-to-record-converter')
